requirements.txt
Django==4.1.4
djangorestframework==3.14.0
drf-yasg==1.21.4
psycopg2-binary==2.9.5

